python/demo2/pyproject.toml
[tool.poetry]
name = "gcp-ml-demo2"
version = "0.1.0"
description = "2nd demo of Google Cloud - Machine Learning Services specialization"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12.0"


[tool.poetry.group.prediction.dependencies]
fastapi = "^0.110.0"
google-cloud-storage = "^2.14.0"
pydantic = "^2.6.2"
uvicorn = "^0.27.1"
pandas = "^2.2.1"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.training.dependencies]
fire = "^0.5.0"
pandas = "^2.2.1"
pyarrow = "^15.0.0"
scikit-learn = "^1.4.1.post1"


[tool.poetry.group.pipeline.dependencies]
kfp = "^2.7.0"
google-cloud-aiplatform = "^1.42.1"
google-cloud-pipeline-components = "^2.10.0"
matplotlib = "^3.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
